requirements.txt
fastapi==0.121.3
SQLAlchemy==2.0.44
uvicorn==0.38.0
requests==2.32.5
httpx==0.28.1
black==25.12.0
pylint==4.0.4
pytest==9.0.2
pytest-cov==7.0.0
python-multipart==0.0.20
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.5.0
fire==0.7.1